backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fiesta-backend"
version = "0.1.0"
description = "FIESTA Backend API for EarthRef.org"
authors = [
    { name = "EarthRef", email = "webmaster@earthref.org" },
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.28.0",
    "elasticsearch>=8.10.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = []

[tool.ruff.isort]
known-first-party = ["app"]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[project.urls]
"Homepage" = "https://earthref.org"
"Bug Tracker" = "https://github.com/earthref/FIESTA/issues"
